% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_gadget}
\alias{translate_gadget}
\alias{translate_langs}
\title{Translate text with a shiny gadget.}
\usage{
translate_gadget(text = "", to = NULL)

translate_langs()
}
\arguments{
\item{text}{Default text to translate.}

\item{to}{Default language to translate to. See \code{translate_lang()}
to get a full list of options.}
}
\value{
A length one character vector containing the translated text.
}
\description{
This uses the MS translation API - you need your key and secret
set as \code{MS_TRANSLATE_KEY} and \code{MS_TRANSLATE_SECRET}
environment variables.
}
\examples{
if (interactive())
  translate_gadget("My name is Hadley", "fr")
}

